requirements.txt
google-genai
httpx
fastapi
uvicorn
